fix: folder structure and layout settings
