+++ src/app/login/components/LoginForm.tsx
@@ -1,16 +1,17 @@
 "use client";
 import React from "react";
 import { FormikProvider, useFormik } from "formik";
+import Image from "next/image";
 import Link from "next/link";
 
-import { Button } from "../../atoms/button";
-import { Input } from "../../atoms/input";
+import { Button } from "@/core/ui/button";
+import { Input } from "@/core/ui/input";
 
-const ResetPassForm = () => {
+const LoginForm = () => {
   const formik = useFormik({
     initialValues: {
-      newPassword: "",
-      confirmPassword: "",
+      email: "",
+      password: "",
     },
 
     onSubmit: (values) => {
@@ -22,31 +23,50 @@ const ResetPassForm = () => {
   return (
     <FormikProvider value={formik}>
       <form onSubmit={formik.handleSubmit} className="flex flex-col gap-4">
-        <div className="flex items-center  border-b border-b-graylight w-[280px]">
+        {/* Email */}
+
+        <div className="flex items-center w-full border-b border-b-graylight ">
           <Input
-            type="password"
-            id="newPassword"
-            name="newPassword"
-            placeholder="New Password"
+            type="email"
+            id="email"
+            name="email"
+            placeholder="Email Address"
             onChange={formik.handleChange}
-            value={formik.values.newPassword}
+            value={formik.values.email}
             className="bg-transparent outline-none  border-none placeholder:text-graylight "
+          />
+          <Image
+            src={"/assets/register_email_icon.svg"}
+            alt="email_icon"
+            width={22}
+            height={22}
           />
         </div>
 
-        {/* Confirm Password */}
+        {/* Password */}
 
         <div className="flex items-center w-full border-b border-b-graylight ">
           <Input
             type="password"
-            id="confirmPassword"
-            name="confirmPassword"
-            placeholder="Confirm Password"
+            id="password"
+            name="password"
+            placeholder="Password"
             onChange={formik.handleChange}
-            value={formik.values.confirmPassword}
+            value={formik.values.password}
             className="bg-transparent outline-none  border-none placeholder:text-graylight "
           />
+          <Image
+            src={"/assets/register_password_icon.svg"}
+            alt="password_icon"
+            width={22}
+            height={22}
+          />
         </div>
+        <Link href="/login/forgot-password">
+          <div className="w-full justify-end text-xs text-white flex cursor-pointer">
+            Forgot Password?
+          </div>
+        </Link>
 
         {/* Submit Button */}
         <div className="flex justify-center items-center mt-2">
@@ -55,7 +75,7 @@ const ResetPassForm = () => {
               className="rounded-full  bg-greentertiary hover:bg-greenaccent text-white  flex justify-center items-center w-36"
               type="submit"
             >
-              Confirm
+              Next
             </Button>
           </Link>
         </div>
@@ -64,4 +84,4 @@ const ResetPassForm = () => {
   );
 };
 
-export default ResetPassForm;
+export default LoginForm;
