+++ src/app/login/verification-code/components/VerifyCodeForm.tsx
@@ -3,8 +3,8 @@ import React from "react";
 import { FormikProvider, useFormik } from "formik";
 import Link from "next/link";
 
-import { Button } from "../../atoms/button";
-import { Input } from "../../atoms/input";
+import { Button } from "@/core/ui/button";
+import { Input } from "@/core/ui/input";
 
 const VerifyCodeForm = () => {
   const formik = useFormik({
