+++ src/core/ui/backButton.tsx
@@ -1,16 +1,16 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import LeftArrowIcon from "../../../../public/assets/left_arrow_icon.svg";
 const BackButton = ({ href }: any) => {
   return (
     <>
       <Link href={href}>
         <div className="flex items-center gap-2 cursor-pointer">
           <Image
-            src={LeftArrowIcon}
+            src={"/assets/left_arrow_icon.svg"}
             alt="left_arrow_icon"
-            className="h-4 w-4 "
+            width={8}
+            height={8}
           />
           <span className="text-white">Back</span>
         </div>
